Moving keys from instruments to consolidated keyboard plus misc. styling adjustments
