+++ src/components/Controls/Controls.css
@@ -1,11 +1,13 @@
 .controls {
     display: flex;
-    flex-wrap: wrap;
     justify-content: center;
+    flex-grow: 0;
+    flex-shrink: 0;
     align-items: center;
     width: 100%;
     font-size: 16px;
     background-color: #111;
+    overflow-x: hidden;
 }
 
 .controls button {
@@ -13,6 +15,10 @@
     margin: 5px 2px;
     background-color: rgb(103, 103, 103);
     color: white;
+}
+
+.controls .button-clear {
+  width: 170px;
 }
 
 .controls-section {
@@ -63,4 +69,11 @@
     border: none;
     width: 70px;
     padding-left: 10px;
+}
+
+@media (max-width: 900px) {
+  .controls {
+    justify-content: flex-start;
+    overflow-x: scroll;
+  }
 }
