+++ src/components/DrumPad/DrumPad.css
@@ -1,9 +1,8 @@
 .drum-pad {
-    font-size: 24px;
-    width: 65px;
-    height: 65px;
+    height: 56px;
+    width: 56px;
     box-sizing: border-box;
-    margin: 5px;
+    margin: 4px;
     border-radius: 10px;
     outline: none;
     user-select: none;
