+++ src/components/Instrument/Instrument.css
@@ -2,22 +2,22 @@
     display: flex;
     flex-direction: column;
     justify-content: flex-start;
+    flex-shrink: 0;
     align-items: center;
-    width: 300px;
-    height: 531px;
+    width: 260px;
     margin: 10px;
-    padding: 40px;
+    padding: 35px;
     background-color: rgba(38, 38, 38, 1);
-    border-radius: 20px;
+    border-radius: 15px;
   }
 
   .display-title {
-    font-size: 24px;
+    font-size: 22px;
     margin: 3px;
   }
 
   .display-content {
-    font-size: 20px;
+    font-size: 18px;
     margin: 3px;
   }
 
@@ -29,19 +29,11 @@
     justify-content: center;
     align-items: center;
     width: 100%;
-    height: 110px;
+    height: 90px;
     box-sizing: border-box;
     text-align: center;
     border-radius: 10px;
     margin-bottom: 20px;
     color: rgb(230, 230, 230);
     background-color: rgba(255, 255, 255, 0.1);
-  }
-  
-  .pad-bank {
-    width: 100%;
-    display: flex;
-    justify-content: flex-start;
-    flex-wrap: wrap;
-    margin-top: 20px;
   }
