--- src/config/keyMappings.js
@@ -1,580 +1,0 @@
-export default {
-  drums: [
-    {
-        id: 1,
-        name: 'High Tom',
-        triggerKey: 'q',
-        source: 'sounds/drums/Tom High/Tom High 4.mp3',
-        volume: 1.0,
-    }, 
-    {
-        id: 2,
-        name: 'Mid Tom',
-        triggerKey: 'w',
-        source: 'sounds/drums/Tom Mid/Tom Mid 4.mp3',
-        volume: 1.0
-    }, 
-    {
-        id: 3,
-        name: 'Floor Tom',
-        triggerKey: 'e',
-        source: 'sounds/drums/Tom Floor/Tom Floor 5.mp3',
-        volume: 1.0
-    }, 
-    {
-        id: 4,
-        name: 'Crash',
-        triggerKey: 'a',
-        source: 'sounds/drums/Crash A/Crash A 5.mp3',
-        volume: 0.55
-    }, 
-    {
-        id: 5,
-        name: 'Ride',
-        triggerKey: 's',
-        source: 'sounds/drums/Ride Cymbal/Ride Cymbal 3.mp3',
-        volume: 0.8
-    }, 
-    {
-        id: 6,
-        name: 'Hi-Hat Open',
-        triggerKey: 'd',
-        source: 'sounds/drums/Hi Hat/Hi Hat Open 5.mp3',
-        volume: 0.5,
-        exclusiveZone: 2
-    }, 
-    {
-        id: 7,
-        name: 'Kick',
-        triggerKey: 'z',
-        source: 'sounds/drums2/kick-softy.mp3',
-        volume: 1
-    }, 
-    {
-        id: 8,
-        name: 'Snare',
-        triggerKey: 'x',
-        source: 'sounds/drums/Snare On/Snare 6.mp3',
-        volume: 0.7
-    }, 
-    {
-        id: 9,
-        name: 'Hi-Hat Closed',
-        triggerKey: 'c',
-        source: 'sounds/drums/Hi Hat/Hi Hat 8.mp3',
-        volume: 0.77,
-        exclusiveZone: 2
-    },
-  ],
-  synth: [
-    {
-        id: 1,
-        name: 'Synth C Major',
-        triggerKey: '4',
-        source: 'sounds/synth/I_Do_Remember_C_01.wav',
-        volume: 0.75,
-        pan: 0.5,
-        //exclusiveZone: 3
-    },
-    {
-        id: 2,
-        name: 'Synth D Major',
-        triggerKey: '5',
-        source: 'sounds/synth/I_Do_Remember_D_01.wav',
-        volume: 0.75,
-        pan: 0.5,
-        //exclusiveZone: 3
-    },  
-    {
-        id: 3,
-        name: 'Synth E Major',
-        triggerKey: '6',
-        source: 'sounds/synth/I_Do_Remember_E_01.wav',
-        volume: 0.75,
-        pan: 0.5,
-        //exclusiveZone: 3
-    },  
-    {
-        id: 4,
-        name: 'Synth F# Major',
-        triggerKey: '7',
-        source: 'sounds/synth/I_Do_Remember_F_01.wav',
-        volume: 0.75,
-        pan: 0.5,
-        //exclusiveZone: 3,
-        detune: 100
-    },  
-    {
-        id: 5,
-        name: 'Synth G Major',
-        triggerKey: '8',
-        source: 'sounds/synth/I_Do_Remember_G_01.wav',
-        volume: 0.75,
-        pan: 0.5,
-        //exclusiveZone: 3
-    },
-    {
-        id: 6,
-        name: 'Synth A Major',
-        triggerKey: '9',
-        source: 'sounds/synth/I_Do_Remember_A_01.wav',
-        volume: 0.75,
-        pan: 0.5,
-        //exclusiveZone: 3
-    },
-    {
-        id: 7,
-        name: 'Synth B Major',
-        triggerKey: '0',
-        source: 'sounds/synth/I_Do_Remember_B_01.wav',
-        volume: 0.75,
-        pan: 0.5,
-        //exclusiveZone: 3
-    },
-  ], 
-  piano: [
-    {
-        id: 1,
-        name: 'Piano Chord 1',
-        triggerKey: 'f',
-        source: 'sounds/piano/Chord_1.mp3',
-        volume: 1.0,
-        pan: -0.5,
-        //exclusiveZone: 1
-    }, 
-    {
-        id: 2,
-        name: 'Piano Chord 2',
-        triggerKey: 'g',
-        source: 'sounds/piano/Chord_2.mp3',
-        volume: 1.0,
-        pan: -0.5,   
-        //exclusiveZone: 1
-    }, 
-    {
-        id: 3,
-        name: 'Piano Chord 3',
-        triggerKey: 'h',
-        source: 'sounds/piano/Chord_3.mp3',
-        volume: 1.0,
-        pan: -0.5,
-        //exclusiveZone: 1
-    }, 
-    {
-        id: 4,
-        name: 'Piano Chord 4',
-        triggerKey: 'j',
-        source: 'sounds/piano/Chord_3.mp3',
-        volume: 1.0,
-        pan: -0.5,
-        //exclusiveZone: 1,
-        detune: 200
-    }, 
-    {
-        id: 5,
-        name: 'Piano Chord 5',
-        triggerKey: 'k',
-        source: 'sounds/piano/Chord_3.mp3',
-        volume: 1.0,
-        pan: -0.5,
-        //exclusiveZone: 1,
-        detune: 300
-    },
-    {
-        id: 6,
-        name: 'Piano Chord 6',
-        triggerKey: 'l',
-        source: 'sounds/piano/Chord_3.mp3',
-        volume: 1.0,
-        pan: -0.5,
-        //exclusiveZone: 1,
-        detune: 500
-    },
-    {
-        id: 7,
-        name: 'Piano Chord 7',
-        triggerKey: ';',
-        source: 'sounds/piano/Chord_3.mp3',
-        volume: 1.0,
-        pan: -0.5,
-        //exclusiveZone: 1,
-        detune: 700
-    },
-    {
-        id: 8,
-        name: 'Piano Chord 7',
-        triggerKey: '\'',
-        source: 'sounds/piano/Chord_3.mp3',
-        volume: 1.0,
-        pan: -0.5,
-        //exclusiveZone: 1,
-        detune: 900
-    }
-  ], 
-  piano2: [
-    {
-        id: 1,
-        name: 'Piano G',
-        triggerKey: '`',
-        source: 'sounds/piano/piano-g_Gsharp_major.wav',
-        volume: 1.0,
-        pan: -0.5,
-        detune: -100
-    },    
-    {
-        id: 2,
-        name: 'Piano G#',
-        triggerKey: '1',
-        source: 'sounds/piano/piano-g_Gsharp_major.wav',
-        volume: 0.98,
-        pan: -0.5,
-    },    
-    {
-        id: 3,
-        name: 'Piano A',
-        triggerKey: '2',
-        source: 'sounds/piano/piano-a_A_major.wav',
-        volume: 0.96,
-        pan: -0.5,
-    },   
-    {
-        name: 'Piano A#',
-        triggerKey: '3',
-        source: 'sounds/piano/piano-bb_Asharp_major.wav',
-        volume: 0.94,
-        pan: -0.5,
-    },   
-    {
-        id: 4,
-        name: 'Piano A#',
-        triggerKey: '4',
-        source: 'sounds/piano/piano-b_B_major.wav',
-        volume: 0.92,
-        pan: -0.5,
-    },      
-    {
-        name: 'Piano C',
-        triggerKey: '5',
-        source: 'sounds/piano/piano-c_C_major.wav',
-        volume: 0.90,
-        pan: -0.5,
-    }, 
-    {
-        id: 5,
-        name: 'Piano C#',
-        triggerKey: '6',
-        source: 'sounds/piano/piano-c_Csharp_major.wav',
-        volume: 0.88,
-        pan: -0.5,
-    }, 
-    {
-        id: 6,
-        name: 'Piano D',
-        triggerKey: '7',
-        source: 'sounds/piano/piano-d_D_major.wav',
-        volume: 0.86,
-        pan: -0.5,
-    }, 
-    {
-        id: 7,
-        name: 'Piano D#',
-        triggerKey: '8',
-        source: 'sounds/piano/piano-eb_Dsharp_major.wav',
-        volume: 0.84,
-        pan: -0.5,
-    }, 
-    {
-        id: 8,
-        name: 'Piano E',
-        triggerKey: '9',
-        source: 'sounds/piano/piano-e_E_major.wav',
-        volume: 0.82,
-        pan: -0.5,
-    }, 
-    {
-        id: 9,
-        name: 'Piano F',
-        triggerKey: '0',
-        source: 'sounds/piano/piano-f_F_major.wav',
-        volume: 0.80,
-        pan: -0.5,
-    },     
-    {
-        id: 10,
-        name: 'Piano F#',
-        triggerKey: '-',
-        source: 'sounds/piano/piano-f_Fsharp_major.wav',
-        volume: 0.78,
-        pan: -0.5,
-    },     
-    {
-        id: 11,
-        name: 'Piano G',
-        triggerKey: '=',
-        source: 'sounds/piano/piano-g_G_major.wav',
-        volume: 0.76,
-        pan: -0.5,
-    },    
-  ],
-  cello: [
-    /*
-    {
-        id: 1,
-        name: 'Cello G',
-        triggerKey: '1',
-        source: 'sounds/strings/cello_G_major.wav',
-        volume: 1.0,
-        detune: 0
-    }, 
-    {
-        id: 2,
-        name: 'Cello A',
-        triggerKey: '2',
-        source: 'sounds/strings/cello_G_major.wav',
-        volume: 0.98,
-        detune: 200
-    }, 
-    {
-        id: 3,
-        name: 'Cello B',
-        triggerKey: '3',
-        source: 'sounds/strings/cello_G_major.wav',
-        volume: 0.96,
-        detune: 400
-    }, 
-    */
-    {
-        id: 4,
-        name: 'Cello C',
-        triggerKey: 'r',
-        source: 'sounds/strings/cello_G_major.wav',
-        volume: 0.94,
-        detune: 500
-    }, 
-    {
-        id: 5,
-        name: 'Cello D',
-        triggerKey: 't',
-        source: 'sounds/strings/cello_G_major.wav',
-        volume: 0.92,
-        detune: 700
-    }, 
-    {
-        id: 6,
-        name: 'Cello E',
-        triggerKey: 'y',
-        source: 'sounds/strings/cello_G_major.wav',
-        volume: 0.90,
-        detune: 900
-    }, 
-    {
-        id: 7,
-        name: 'Cello F#',
-        triggerKey: 'u',
-        source: 'sounds/strings/cello_G_major.wav',
-        volume: 0.88,
-        detune: 1100
-    }, 
-    {
-        id: 8,
-        name: 'Cello G',
-        triggerKey: 'i',
-        source: 'sounds/strings/cello_G_major.wav',
-        volume: 0.86,
-        detune: 1200
-    }, 
-    {
-        id: 9,
-        name: 'Cello A',
-        triggerKey: 'o',
-        source: 'sounds/strings/cello_G_major.wav',
-        volume: 0.84,
-        detune: 1400
-    }, 
-    {
-        id: 10,
-        name: 'Cello B',
-        triggerKey: 'p',
-        source: 'sounds/strings/cello_G_major.wav',
-        volume: 0.82,
-        detune: 1600
-    }, 
-    {
-        id: 11,
-        name: 'Cello C',
-        triggerKey: '[',
-        source: 'sounds/strings/cello_G_major.wav',
-        volume: 0.80,
-        detune: 1700
-    }, 
-    /*
-    {
-        id: 12,
-        name: 'Cello D',
-        triggerKey: ']',
-        source: 'sounds/strings/cello_G_major.wav',
-        volume: 0.78,
-        detune: 1900
-    }, 
-    */
-  ],
-  synthDrums: [
-    {
-        id: 1,
-        name: '808 C',
-        triggerKey: 'v',
-        source: 'sounds/drums2/808-bass-f-raw_Fsharp.wav',
-        volume: 1.0,
-        detune: -600,
-        //exclusiveZone: 6
-    }, 
-    {
-        id: 2,
-        name: '808 D',
-        triggerKey: 'b',
-        source: 'sounds/drums2/808-bass-f-raw_Fsharp.wav',
-        volume: 1.0,
-        detune: -400,
-        //exclusiveZone: 6
-    }, 
-    {
-        id: 3,
-        name: '808 E',
-        triggerKey: 'n',
-        source: 'sounds/drums2/808-bass-f-raw_Fsharp.wav',
-        volume: 1.0,
-        detune: -200,
-        //exclusiveZone: 6
-    }, {
-        id: 4,
-        name: '808 F#',
-        triggerKey: 'm',
-        source: 'sounds/drums2/808-bass-f-raw_Fsharp.wav',
-        volume: 1.0,
-        detune: 0,
-        //exclusiveZone: 6
-    }, {
-        id: 5,
-        name: '808 G',
-        triggerKey: ',',
-        source: 'sounds/drums2/808-bass-f-raw_Fsharp.wav',
-        volume: 1.0,
-        detune: 100,
-        //exclusiveZone: 6
-    }, {
-        id: 6,
-        name: '808 A',
-        triggerKey: '.',
-        source: 'sounds/drums2/808-bass-f-raw_Fsharp.wav',
-        volume: 1.0,
-        detune: 300,
-        //exclusiveZone: 6
-    }, {
-        id: 7,
-        name: '808 B',
-        triggerKey: '/',
-        source: 'sounds/drums2/808-bass-f-raw_Fsharp.wav',
-        volume: 1.0,
-        detune: 500,
-        //exclusiveZone: 6
-    }
-  ]
-}
-
-    /*
-    {
-        name: 'Synth Choir Chord',
-        triggerKey: ',',
-        source: 'sounds/synth/choir-chord-pad_78bpm_B_major.wav',
-        volume: 1.0,
-        //exclusiveZone: 5
-    }, 
-    
-
-    {
-        name: 'Cinematic Strings',
-        triggerKey: 'n',
-        source: 'sounds/strings/cinematic-strings_125bpm_G_minor.wav',
-        volume: 1.0,
-        //exclusiveZone: 5
-    }, 
-    {
-        name: 'Cinematic Strings',
-        triggerKey: 'm',
-        source: 'sounds/strings/violin-arp_128bpm_Dsharp_minor.wav',
-        volume: 1.0,
-        //exclusiveZone: 5
-    }, 
-    */
-
-    /*
-export const synth2 = [
-    {
-        name: 'Madmax Pad C',
-        triggerKey: 'v',
-        source: 'sounds/synth/Madmax_Pad_E.wav',
-        volume: 0.75,
-        pan: 0.5,
-        //exclusiveZone: 4,
-        detune: -400
-    },
-    {
-        name: 'Madmax Pad D',
-        triggerKey: 'b',
-        source: 'sounds/synth/Madmax_Pad_E.wav',
-        volume: 0.75,
-        pan: 0.5,
-        //exclusiveZone: 4,
-        detune: -200
-    },
-    {
-        name: 'Madmax Pad E',
-        triggerKey: 'n',
-        source: 'sounds/synth/Madmax_Pad_E.wav',
-        volume: 0.75,
-        pan: 0.5,
-        //exclusiveZone: 4,
-        detune: 0
-    },
-    {
-        name: 'Orion Pad F#',
-        triggerKey: 'm',
-        source: 'sounds/synth/Orion_Pad.wav',
-        volume: 0.75,
-        pan: 0.5,
-        //exclusiveZone: 4,
-        detune: 600
-    },
-    {
-        name: 'Orion Pad G',
-        triggerKey: ',',
-        source: 'sounds/synth/Orion_Pad.wav',
-        volume: 0.75,
-        pan: 0.5,
-        //exclusiveZone: 4,
-        detune: 700
-    },
-    {
-        name: 'Orion Pad A',
-        triggerKey: '.',
-        source: 'sounds/synth/Orion_Pad.wav',
-        volume: 0.75,
-        pan: 0.5,
-        //exclusiveZone: 4,
-        detune: 900
-    },
-    {
-        name: 'Orion Pad B',
-        triggerKey: '/',
-        source: 'sounds/synth/Orion_Pad.wav',
-        volume: 0.75,
-        pan: 0.5,
-        //exclusiveZone: 4,
-        detune: 1100
-    },
-    */
-
-    
-    
-    
-
