+++ src/config/keyboards.js
@@ -1,0 +1,19 @@
+export default {
+  mac: [
+    [
+      ['~'], ['1'], ['2'], ['3'], ['4'], ['5'], ['6'], ['7'], ['8'], ['9'], ['0'], ['-'], ['='], ['delete', 1.68, 's']
+    ],
+    [
+      ['tab', 1.68, 's'], ['q'], ['w'], ['e'], ['r'], ['t'], ['y'], ['u'], ['i'], ['o'], ['p'], ['['], [']'], ['\\']
+    ],
+    [
+      ['caps lock', 1.9, 's'], ['a'], ['s'], ['d'], ['f'], ['g'], ['h'], ['j'], ['k'], ['l'], [';'], ['\''], ['return', 1.9, 's']
+    ],
+    [
+      ['shift', 2.45, 's'], ['z'], ['x'], ['c'], ['v'], ['b'], ['n'], ['m'], [','], ['.'], ['/'], ['shift', 2.45, 's']
+    ],
+    [
+      ['fn', 1, 's'], ['ctrl', 1, 's'], ['opt', 1, 's'], ['cmd', 1, 's'], ['', 6], ['cmd', 1, 's'], ['opt', 1, 's']
+    ]
+  ]
+}
