+++ src/index.css
@@ -24,3 +24,17 @@ button {
   font-size: 16px;
   outline: none;
 }
+
+::-webkit-scrollbar {
+  background: rgba(25, 25, 25, 0.7);
+}
+
+::-webkit-scrollbar-thumb {
+  background: rgb(95, 95, 95, 0.7);
+  -webkit-border-radius: 1ex;
+  -webkit-box-shadow: 0px 1px 2px rgba(25, 25, 25, 0.7);
+}
+
+::-webkit-scrollbar-corner {
+  background: rgba(25, 25, 25, 0.7);
+}
