+++ src/reducers/index.js
@@ -1,8 +1,10 @@
 import { combineReducers } from 'redux';
 import record from './record';
 import controls from './controls';
+import instruments from './instruments';
 
 export default combineReducers({
     record,
-    controls
+    controls,
+    instruments
 });
