+++ src/reducers/instruments.js
@@ -1,0 +1,5 @@
+import initialState from '../config/instruments';
+
+export default (state = initialState, action) => {
+  return state;
+}
