+++ src/services/detectBrowser.js
@@ -1,0 +1,21 @@
+export const detectBrowser = () => { 
+  let isMobile = false;
+  let isChrome = false;
+  if (
+    navigator.userAgent.match(/Android/i) ||
+    navigator.userAgent.match(/webOS/i) ||
+    navigator.userAgent.match(/iPhone/i) ||
+    navigator.userAgent.match(/iPad/i) ||
+    navigator.userAgent.match(/iPod/i) ||
+    navigator.userAgent.match(/BlackBerry/i) ||
+    navigator.userAgent.match(/Windows Phone/i)
+  ) {
+    isMobile = true;
+  }
+    
+  if (navigator.userAgent.match(/Chrome/)) {
+    isChrome = true;
+  }
+
+  return { isMobile, isChrome };
+}
