+++ src/services/index.js
@@ -1,4 +1,4 @@
-export * from './detectMobile.js';
+export * from './detectBrowser.js';
 export * from './handleKeyEvent';
 export * from './loadAudioBuffer';
 export * from './reverbs';
